Extracted token utils to seperate files
