+++ inc/tokenisation.h
@@ -10,23 +10,23 @@ typedef enum s_bool
 }	t_bool;
 
 typedef enum s_token_type {
-    TOKEN_WORD,      // General word or command
-    TOKEN_AND,       // &&
-    TOKEN_OR,        // ||
-    TOKEN_LPAREN,    // (
-    TOKEN_RPAREN,    // )
-    TOKEN_PIPE,      // |
-    TOKEN_REDIRECT_IN,   // <
-    TOKEN_REDIRECT_OUT,  // >
-    TOKEN_APPEND,        // >>
-    TOKEN_HEREDOC,       // <<
-    // ... Add other tokens as needed
+	TOKEN_WORD,      // General word or command
+	TOKEN_AND,       // &&
+	TOKEN_OR,        // ||
+	TOKEN_LPAREN,    // (
+	TOKEN_RPAREN,    // )
+	TOKEN_PIPE,      // |
+	TOKEN_REDIRECT_IN,   // <
+	TOKEN_REDIRECT_OUT,  // >
+	TOKEN_APPEND,        // >>
+	TOKEN_HEREDOC,       // <<
+	// ... Add other tokens as needed
 }   t_token_type;
 
 typedef struct s_token
 {
-    t_token_type    type;
-    char            *value;
+	t_token_type    type;
+	char            *value;
 }   t_token;
 
 typedef struct s_token_node
@@ -41,9 +41,13 @@ typedef struct s_token_list
 	t_token_node *tail;
 }   t_token_list;
 
+t_token_list    *init_token_list(void);
+void            add_token(t_token_list *list, t_token *token);
+t_bool          ft_is_whitespace(char c);
 t_token         *create_token(t_token_type type, char *value);
 t_token_list	*ft_tokenize(const char *input);
 void            free_token_list(t_token_list *list);
 void            print_token_list(t_token_list *list);
+void            free_token(t_token *token);
 
 #endif
