+++ src/parsing/tokenisation/tokens.c
@@ -1,98 +1,5 @@
 #include <tokenisation.h>
 
-t_bool	ft_is_whitespace(char c)
-{
-	return (c == ' ' || c == '\t');
-}
-
-t_token *create_token(t_token_type type, char *value)
-{
-	t_token *token;
-	
-	token = malloc(sizeof(t_token));
-	token->type = type;
-	if (value)
-		token->value = ft_strdup(value);
-	else 
-		token->value = NULL;
-	return (token);
-}
-
-static void    free_token(t_token *token)
-{
-	if (token)
-	{
-		if (token->value)
-			free(token->value);
-		free(token);
-	}
-}
-
-static t_token_list    *init_token_list(void)
-{
-	t_token_list    *list;
-
-	list = malloc(sizeof(t_token_list));
-	list->head = NULL;
-	list->tail = NULL;
-	return (list);
-}
-
-void    free_token_list(t_token_list *list)
-{
-	t_token_node    *current;
-	t_token_node    *next;
-
-	if (list == NULL)
-		return ;
-	current = list->head;
-	while (current)
-	{
-		next = current->next;
-		free_token(current->token);
-		free(current);
-		current = next;
-	}
-	free(list);
-}
-
-static void    add_token(t_token_list *list, t_token *token)
-{
-	t_token_node    *node;
-
-	node = malloc(sizeof(t_token_node));
-	node->token = token;
-	node->next = NULL;
-	if (list->tail)
-	{
-		list->tail->next = node;
-		list->tail = node;
-	}
-	else
-	{
-		list->head = node;
-		list->tail = node;
-	}
-}
-
-void print_token_list(t_token_list *list)
-{
-	t_token_node    *current;
-
-	if (list == NULL)
-		return ;
-	current = list->head;
-	while (current != NULL)
-	{
-		printf("Token: (%d)", current->token->type);
-		if (current->token->value != NULL)
-			printf(" (%s)\n", current->token->value);
-		else
-			printf("\n");
-		current = current->next;
-	}
-}
-
 t_token_list	*ft_tokenize(const char *input)
 {
 	const char		*current;
